requirements.txt
google-adk==1.25.1
google-genai==1.64.0
google-cloud-aiplatform>=1.70.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.24.0